single player bug fixes
