+++ Intersect.Server.Core/Database/DataMigrations/Sqlite/SqliteUserVariablePopulateNewColumnId.cs
@@ -1,0 +1,7 @@
+namespace Intersect.Server.Database.DataMigrations.Sqlite;
+
+[SchemaMigration(typeof(Migrations.Sqlite.Player.UserVariables_AddStandaloneIdPK), ApplyIfLast = true)]
+public sealed class SqliteUserVariablePopulateNewColumnId : UserVariablePopulateNewColumnId
+{
+
+}
